requirements.txt
numpy==1.25.2
tensorflow==2.14.0
uvicorn==0.29.0
fastapi==0.110.2
Pillow==9.3.0